+++ test/test_ops.py
@@ -133,6 +133,7 @@ class TestOps(unittest.TestCase):
     helper_test_op([(), ()], torch.minimum, Tensor.minimum)
   def test_add(self):
     helper_test_op([(45,65), (45,65)], lambda x,y: x+y, Tensor.add)
+  def test_add_number(self):
     helper_test_op([(), ()], lambda x,y: x+y, Tensor.add)
   def test_add3(self):
     helper_test_op([(45,65), (45,65), (45,65)], lambda x,y,z: x+y+z)
